% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_week.R
\name{calculate_week}
\alias{calculate_week}
\title{Calculate the number of weeks past a given date}
\usage{
calculate_week(
  x,
  start_date,
  no_of_weeks = 1,
  date_parsing = lubridate::ymd,
  as_integer = TRUE
)
}
\arguments{
\item{x}{A character vector of suspected dates. The dates should be in the parse format specified in \code{date_parsing}. No default value given.}

\item{start_date}{A single reference date which week 0 starts from. No default value given.}

\item{no_of_weeks}{The number of weeks to consider as a single unit. For example, this can be used to identify bi-weeks or any integer multiplication of weeks. The default is given as 1, which gives the number of weeks. For the bi-week use \code{no_of_weeks = 2}.}

\item{date_parsing}{A function for parsing the date from the \code{lubridate} package. The default is the \code{ymd} year-month-date format.}

\item{as_integer}{Specifies if the output should be an integer vector of weeks (default). If \code{FALSE} will output a tibble with the input set of dates and the weeks.}
}
\value{
If \code{as_integer = TRUE}: a numeric vector of the same length as \code{x} which contains the week number (default), or any other number of week increments specified in \code{no_of_weeks}. The week can be used in other calculations or models as necessary. Negative values indicates the weeks before the starting date. If \code{as_integer = FLASE}: a tibble with the input dates and the week.
}
\description{
This function calculates the number of weeks past a given "start date" for a vector of dates. This function also allows to combine multiple weeks into a single unit and calculate the "week" (e.g.: bi-weekly, etc.). One use of this function is data aggregation (e.g.: convert daily data into weekly data).
}
\examples{
d <- c("2020-01-01", "2020-02-01")
calculate_week(d, start_date = "2020-01-2")
}
